third_party/blink/web_tests/wpt_internal/css/css-contain/container-queries/at-container-snapped-serialization.html

<!DOCTYPE html>
<title>CSS Container Queries: state(snapped) conditionText serialization</title>
<link rel="help" href="https://drafts.csswg.org/css-contain-3/#container-queries">
<link rel="help" href="https://drafts.csswg.org/cssom/#serialize-a-css-rule">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/css-contain/container-queries/support/cq-testcommon.js"></script>
<style id="testSheet">
  @container state(        snapped:inline) { }
  @container STate(snapped:    ) { }
  @container STate(snapped) { }
  @container  state(  ( snapped: INLINE) OR ( SNAPPED: BlOCK  ) ) { }
  @container state (snapped: inline) { }
</style>
<script>
  setup(() => {
    assert_implements_container_queries();
    assert_equals(testSheet.sheet.cssRules.length, 5);
  });

  const tests = [
      ["state(snapped: inline)", "Normalize spaces"],
      ["STate(snapped: )", "No value - invalid, serializes as <general-enclosed>"],
      ["state(snapped)", "Boolean context"],
      ["state((snapped: inline) or (snapped: block))", "Logical with 'or'"],
      ["state (snapped: inline)", "Not a state function with space before '('"]
  ].map((e, i) => [testSheet.sheet.cssRules[i], ...e]);

  tests.forEach((t) => {
    test(() => assert_equals(t[0].conditionText, t[1]), t[2]);
  });
</script>
